Fixed recompiling not destroying the old behavior
Changed the Evaluator to be a normal instance variable instead of static
to prevent a bug where recompiling the code would not destroy the
original attached behavior.

Added a rotation script and cube to make a better visual example.

Changed Unity to run in the background so you can see script changes
while still inside the IDE.